requirements.txt
playwright
chromadb
flask
google-generativeai
python-dotenv

